fixed an issue in ctx in read cluster
